move input fn into input module
